requirements.txt
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.2.25
charset-normalizer==3.4.6
dnspython==2.8.0
greenlet==3.3.2
idna==3.11
numpy==2.4.3
pandas==3.0.1
playwright==1.58.0
pyee==13.0.1
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
